% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSimLab.R
\name{addPostHook}
\alias{addPostHook}
\title{Hook for functions that are executed after the simulation.}
\usage{
addPostHook(rSimLab, cond = TRUE, ...)
}
\arguments{
\item{rSimLab}{object of class rSimLab}

\item{cond}{condition, that must hold TRUE for the modification
to be executed}

\item{...}{Name-value pairs of expressions. Use NULL to drop a variable.}
}
\value{
object of class rSimLab
}
\description{
Functions are executed row by row. At first, variables are evaluated
in the environment of the calling function, then in the environment of
the parameters and settings.
}
