% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSimLab.R
\name{addPraeHook}
\alias{addPraeHook}
\title{Hook for functions that are executed before the simulation.}
\usage{
addPraeHook(rSimLab, paramToModify, modification, cond = TRUE,
  addEnv = environment())
}
\arguments{
\item{rSimLab}{object of class rSimLab}

\item{paramToModify}{name of the parameter that should
be modified}

\item{modification}{formular for modification}

\item{cond}{condition, that must hold TRUE for the modification
to be executed}

\item{addEnv}{additional environment for the modification}
}
\value{
object of class rSimLab
}
\description{
Hook for functions that are executed before the simulation.
}
\examples{
#' # makes the code a lot more readable
library(magrittr)
analyte(data.frame('time' = 1:10, 'type' = 'pat')) \%>\%
  ana_distrNorm(center = 5, spread = 3) \%>\%
  ana_addQC(5, 'highQC') \%>\%
  addPraeHook('center',
               center + 5,
               type != 'highQC')
}

