% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simAnalyte.R
\name{ana_addQC}
\alias{ana_addQC}
\title{Add Quality Control (QC) measurement}
\usage{
ana_addQC(ana, value, type = "qc")
}
\arguments{
\item{ana}{(unevaluated) object of class analyte}

\item{value}{true value of the QC measurement}

\item{type}{identifier of the QC measurement}
}
\value{
(unevaluated) object of class analyte
}
\description{
This function adds QC measurements with known true values.
}
\details{
One QC measurement is added to each 'time' periode. Therefore
the time variable needs to be present in settings. The type
variable is used to differentiate between patient and QC
measurements and needs to be present in settings as well.
QC measurements are added after everthing else is done
(postHook).
}
\seealso{
Other analyte.blocks: \code{\link{ana_distrLnorm}}
}

