% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simAnalyte.R
\name{ana_distrLnorm}
\alias{ana_distrLnorm}
\alias{ana_distrNorm}
\title{Distribution of true values}
\usage{
ana_distrLnorm(ana, center, spread)

ana_distrNorm(ana, center, spread)
}
\arguments{
\item{ana}{(unevaluated) object of class analyte}

\item{center}{center of distribution}

\item{spread}{spread of distribution}
}
\value{
(unevaluated) object of class analyte
}
\description{
Creates log normal or normal distribution of values. Spread
and center are the real center and spread not
(log-) tranformed in any way for any distribution.
}
\references{
Haeckel, Rainer, and Werner Wosniok.
   "Observed, unknown distributions of clinical chemical
   quantities should be considered to be log-normal: a
   proposal." Clinical chemistry and laboratory
   medicine 48.10 (2010): 1393-1396.
}
\seealso{
Other analyte.blocks: \code{\link{ana_addQC}}
}

