% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simAnalyte.R
\name{analyte}
\alias{analyte}
\title{Analyte}
\usage{
analyte(
  settings = data.frame(),
  params = data.frame(resultName = "trueValue"),
  praeHook = list(),
  postHook = list()
)
}
\arguments{
\item{settings}{a data.frame with settings for analytes}

\item{params}{a data.fram with additional params}

\item{praeHook}{list of functions to modify settings and params}

\item{postHook}{list of functions to modify analytes}
}
\value{
object of class analyte
}
\description{
Creates analytes from settings.
}
\details{
The function analyte uses settings and additional params to
create true values for an analyte. First, settings and params
need to be supplied. The functions in the praeHook list
are executed to modify params. Settings and params are in
scope of these functions directly. The sim.f function creates
true values for the analyte. The functions in the postHook list
are executed to modify these true values. Again, settings and
params are in scope of these functions directly.
}
\seealso{
Other mainSteps: 
\code{\link{measurement}()}
}
\concept{mainSteps}
