% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simMeasurement.R
\name{measurement}
\alias{measurement}
\title{Measurements}
\usage{
measurement(
  settings = data.frame(),
  params = data.frame(measuredName = "trueValue", resultName = "measurement"),
  praeHook = list(),
  postHook = list()
)
}
\arguments{
\item{settings}{data.frame with 'trueValue colume' or analyte
object}

\item{params}{a data.fram with additional params}

\item{praeHook}{list of functions to modify settings and params}

\item{postHook}{list of functions to modify analytes}
}
\value{
measurement object
}
\description{
The function measurement uses settings and additional params to
simulates measurements from true values of analytes. First,
settings and params need to be supplied. The functions in
the praeHook list are executed to modify params. Settings
and params are in scope of these functions directly. The
sim.fxs functions simulate measurements from true values.
The functions in the postHook list are executed to
modify these measuerments. Again, settings and
params are in scope of these functions directly.
}
\seealso{
Other mainSteps: 
\code{\link{analyte}()}
}
\concept{mainSteps}
