% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simMeasurement.R
\name{mm_accFunc}
\alias{mm_accFunc}
\title{Simulates constant and relative accuracy.}
\usage{
mm_accFunc(measurement, constD = 0, relD = 0)
}
\arguments{
\item{measurement}{(unevaluated) object of class measurement}

\item{constD}{constant deviation}

\item{relD}{relative deviation}
}
\value{
(unevaluated) object of class measurement
}
\description{
Simulates constant and relative accuracy.
}
\seealso{
Other measurement.blocks: \code{\link{mm_precCharFunc}}
}

