% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simMeasurement.R
\name{mm_lotBiasFunc}
\alias{mm_lotBiasFunc}
\title{Adds a bias from a reagent lot}
\usage{
mm_lotBiasFunc(measurement, relDLot = 0)
}
\arguments{
\item{measurement}{(unevaluated) object of class measurement}

\item{relDLot}{relative deviation attributable  to a reagent lot}
}
\value{
(unevaluated) object of class measurement
}
\description{
Adds a bias from a reagent lot
}
\seealso{
Other measurement blocks: 
\code{\link{mm_accFunc}()},
\code{\link{mm_precCharFunc}()}
}
\concept{measurement blocks}
