% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simMeasurement.R
\name{mm_lotBiasVariationFunc}
\alias{mm_lotBiasVariationFunc}
\title{Adds reagent lots with different bias. Distribution of lot biases can be
specified as normal distribution}
\usage{
mm_lotBiasVariationFunc(
  measurement,
  sdRelDLot,
  systemicRelDLot = 0,
  nTimesPerLot = NULL
)
}
\arguments{
\item{measurement}{(unevaluated) object of class measurement}

\item{sdRelDLot}{standard deviation of distribution of lot biases}

\item{systemicRelDLot}{mean of distribution of lot biases; systemic bias}

\item{nTimesPerLot}{times (e.g. days) each lot is in use}
}
\value{
(unevaluated) object of class measurement
}
\description{
Adds reagent lots with different bias. Distribution of lot biases can be
specified as normal distribution
}
\seealso{
Other measurement blocks: 
\code{\link{mm_accFunc}()},
\code{\link{mm_lotBiasFunc}()},
\code{\link{mm_precCharFunc}()}
}
\concept{measurement blocks}
