% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simMeasurement.R
\name{mm_accFunc}
\alias{mm_accFunc}
\alias{mm_truenessFunc}
\title{Simulates constant and relative trueness.}
\usage{
mm_accFunc(measurement, constD = 0, relD = 0)

mm_truenessFunc(measurement, constD = 0, relD = 0)
}
\arguments{
\item{measurement}{(unevaluated) object of class measurement}

\item{constD}{constant deviation}

\item{relD}{relative deviation}
}
\value{
(unevaluated) object of class measurement
}
\description{
Systematic errors lead to a (constant or relativ) bias.
The performance characteristics measuring bias is called "trueness".
"Accuracy" is an outdated term for the same concept. See the
reference for more details.
}
\references{
Menditto, Antonio, Marina Patriarca, and Bertil Magnusson.
"Understanding the meaning of accuracy, trueness and precision."
 Accreditation and quality assurance 12.1 (2007): 45-47.
}
\seealso{
Other measurement blocks: 
\code{\link{mm_lotBiasFunc}()},
\code{\link{mm_lotBiasVariationFunc}()},
\code{\link{mm_precCharFunc}()}
}
\concept{measurement blocks}
