% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSimLab.R
\docType{package}
\name{rSimLab}
\alias{rSimLab}
\alias{rSimLab-package}
\title{rSimLab: A package for the simulation of measurements
in laboratory medicine.}
\description{
The rSimLab package simulates measurements in laboratory
medicine. The general idea is to first create analytes with
true values. Measured values are based on these true values.
Complex settings can be created from simple building
blocks. Hooks allow for even more flexibility. The '\%>\%'
operator from the magrittr package is strongly recomended for
chaining everything together.
}
\examples{
# makes the code a lot more readable
library(magrittr)
# Simulation of one year of albumin measurements
setting <- simpleOverTime(365,500, 50)
ana <- analyte(setting) \%>\%
   ana_distrNorm(3.5 , .886)
mm <- measurement(ana) \%>\%
   mm_precCharFunc(.067, 0.031) \%>\%
   mm_accFunc(0, .03)
runSim(mm)

# Simulation of one year of Vitamin D3 with seasonal
# fluctuations.
ana <- analyte(setting) \%>\%
   ana_distrLnorm(19.5, 11.4) \%>\%
   addPraeHook(TRUE, center = center *(sin(time/365*3.14*2)*.075+1))
runSim(ana)

# Repeated measurements of the same sample with different
# methods (Analyte simulation is run explicitly.)
ana <- analyte(setting) \%>\%
   ana_distrLnorm(19.5, 11.4)
ana_concrete <- runSim(ana)
mm1 <- measurement(ana_concrete) \%>\%
   mm_precCharFunc(.05, 0.02)
mm2 <- measurement(ana_concrete) \%>\%
   mm_precCharFunc(.03, 0.1)
runSim(mm1)
runSim(mm2)

# Repeated measurements of the different samples
# (of the same analyte) with different
# methods (Analyte simulation is not run explicitly.)
ana <- analyte(setting) \%>\%
   ana_distrLnorm(19.5, 11.4)
mm1 <- measurement(ana) \%>\%
   mm_precCharFunc(.05, 0.02)
mm2 <- measurement(ana) \%>\%
   mm_precCharFunc(.03, 0.1)
runSim(mm1)
runSim(mm2)
}

