% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSimLab.R, R/simMeasurement.R
\name{runSim}
\alias{runSim}
\alias{runSim.default}
\alias{runSim.measurement}
\title{Executes the simulation}
\usage{
runSim(rSimLab)

\method{runSim}{default}(rSimLab)

\method{runSim}{measurement}(rSimLab)
}
\arguments{
\item{rSimLab}{object of class rSimLab}
}
\value{
data.frame
}
\description{
Executes the simulation
}

