% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simSettings.R
\name{simpleOverTime}
\alias{simpleOverTime}
\title{SimpleOverTime}
\usage{
simpleOverTime(numberOfDays, meanMeasPerDay, sdMeasPerDay)
}
\arguments{
\item{numberOfDays}{number of 'days'}

\item{meanMeasPerDay}{mean number of measurements per 'day'}

\item{sdMeasPerDay}{sd of measurements per 'day'}
}
\value{
data.frame with a 'time' columne
}
\description{
Creates a data.frame with time points for measurements.
}
\examples{
simpleOverTime(10, 4, 2)
}
