% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSimLab.R
\name{addPraeHook}
\alias{addPraeHook}
\title{Hook for functions that are executed before the simulation.}
\usage{
addPraeHook(rSimLab, cond = TRUE, ...)
}
\arguments{
\item{rSimLab}{object of class rSimLab}

\item{cond}{condition, that must hold TRUE for the modification
to be executed}

\item{...}{Name-value pairs of expressions. Use NULL to drop a variable.}
}
\value{
object of class rSimLab
}
\description{
Functions are executed row by row. At first, variables are evaluated
in the environment of the calling function, then in the environment of
the parameters. PraeHook Functions can only modify parameters and
settings.
}
\examples{
# makes the code a lot more readable
library(magrittr)
diff <- 5
analyte(data.frame('time' = 1:10, 'type' = 'pat')) \%>\%
  ana_distrNorm(center = 5, spread = 3) \%>\%
  ana_addQC(5, 'highQC') \%>\%
  addPraeHook(type != 'highQC', center = center + 5) \%>\%
  addPraeHook(type != 'highQC', spread = runif(1,2,3)) \%>\%
  addPraeHook(type != 'highQC', spread = spread + diff)

}
